requirements.txt
autopep8==2.0.1
cachetools==5.3.0
certifi==2022.12.7
chardet==5.1.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==1.2.0
Django==4.1
django-cors-headers==3.13.0
django-timezone-field==5.0
djangorestframework==3.14.0
psycopg2-binary==2.9.5
djangorestframework-api-key==2.*
django-environ==0.9.0
django-seed==0.3.1